flask/pyproject.toml
[project]
name = "Vocabulary_Website_Project"
version = "0.1.0"
description = "A web application for managing vocabulary and language learning."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "lucasw0908" }, 
    { name = "IanWen103026"}
]
requires-python = ">=3.11"
dependencies = [
    "beartype==0.21.0",
    "cryptography>=45.0.7",
    "Flask==3.1.2",
    "Flask-Admin==1.6.1",
    "Flask-Bcrypt==1.0.1",
    "Flask-Babel>=4.0.0",
    "Flask-Migrate==4.0.5",
    "Flask-SQLAlchemy==3.1.1",
    "Flask-Login==0.6.3",
    "Flask-Session==0.8.0",
    "Flask-WTF==1.2.1",
    "google-auth-oauthlib==1.2.2",
    "google-generativeai==0.8.5",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.0.0",
    "wtforms==3.1.2",
    "Zenora==0.0.3.post1",
    "pillow>=11.3.0",
]

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
    "ruff>=0.13.0",
]
